+++ .github/workflows/gitactions.yml
@@ -37,3 +37,4 @@ jobs:
           path: SpaceBattle.Lib.Test/coverage/coverage.opencover.xml     
       - name: Print code coverage
         run: echo "Code coverage percentage ${{steps.create_coverage_badge.outputs.percentage}}%"
+
