+++ SpaceBattle.Lib.Test/AngleTest.cs
@@ -63,5 +63,3 @@ public class AngleTest
     }
 
 };
-
-
