+++ SpaceBattle.Lib.Test/MoveTest.cs
