+++ SpaceBattle.Lib.Test/RotateTest.cs
