+++ SpaceBattle.Lib.Test/VectorTest.cs
