+++ SpaceBattle.Lib/Interfaces/IMovable.cs
