+++ SpaceBattle.Lib/Interfaces/IMoveStopable.cs
