+++ SpaceBattle.Lib/Interfaces/IRotatable.cs
