+++ SpaceBattle.Lib/Interfaces/IUobect.cs
