+++ SpaceBattle.Lib/Interfaces/Iinjecktable.cs
