--- SpaceBattle.Lib/Move/Istartable.cs
@@ -1,11 +1,0 @@
-namespace SpaceBattle.Lib;
-public interface IStartable
-{   IUObject Target
-    {
-        get;
-    }
-    IDictionary<string, object> Properties
-    {
-        get;
-    }
-}
