+++ SpaceBattle.Lib/Strategies/GameOperationMovingStrategy.cs
@@ -1,26 +1,5 @@
 namespace SpaceBattle.Lib;
 using Hwdtech;
-using Hwdtech.Ioc;
-public class StartMoveCommand : ICommand
-{
-    private IStartable startable;
-    public StartMoveCommand(IStartable startable)
-    {
-        this.startable = startable;
-    }
-
-    public void Execute()
-    {
-        startable.Properties.ToList().ForEach(o => IoC.Resolve<ICommand>("Game.Commands.SetProperty", startable.Target, o.Key, o.Value).Execute());
-
-        ICommand cmd = IoC.Resolve<ICommand>("Game.Operations.Moving", startable.Target);
-        IoC.Resolve<ICommand>("Game.Commands.SetProperty", startable.Target, "move", cmd);
-        IoC.Resolve<ICommand>("Game.Queue.Push", IoC.Resolve<Queue<ICommand>>("Game.Queue"), cmd).Execute();
-
-    }
-
-}
-
 
 public class GameOperationsMovingStaregy : IStrategy
 {
@@ -41,6 +20,5 @@ public class GameOperationsMovingStaregy : IStrategy
         list_command.Append(repeatCommand);
 
         return inject_command;
-
     }
 }
